refactor: Phase 2 eliminate pure-wire DTOs and move chain clients to pkg (#614)
* refactor(btc): Phase 2 eliminate pure-wire BTC DTOs (tasks 6.1-6.4)

- Update BTC/BCH port interface signatures to return pkg/chains/btc/rpc
  types directly: GetAddressInfo, ValidateAddress, GetNetworkInfo,
  GetBlockchainInfo, GetTransactionByTxID, DecodeRawTransaction,
  FundRawTransaction, Logging, AddMultisigAddress, ImportDescriptors,
  ImportMulti, GetDescriptorInfo, ListDescriptors
- Update all BTC/BCH adapter implementations accordingly
- Update all BTC/BCH use cases to use btcrpc types
  (monitor_transaction, import_descriptor, import_address,
  create_multisig_address, import_private_key)
- Fix PubKey → Pubkey field rename in psbt_bip32.go and keygen/sign
  import_private_key use cases
- Remove 15 obsolete pkg-to-DTO mapper functions from mapper.go
- Remove 21 pure-wire DTO type definitions from application dto package
  (AddressInfo, NetworkInfo, TransactionResult, DescriptorInfo, etc.)
- Retain domain-enriched DTOs: PreviousTx, ParsedPSBT, UnspentOutput,
  RawTransaction (used by PSBT mapping and signing flows)
- Regenerate MockBitcoiner with updated interface signatures
- All 1142 unit tests pass; lint clean


* refactor(xrp): Phase 2 eliminate pure-wire XRP DTOs and move xrplgo to pkg (tasks 7.1-8.2)

- Update XRP port interfaces to return *xrpclient.* types (SentTx, TxInfo, AccountInfo)
- Update XRP infrastructure adapter (xrpapi_tx.go) with conversion helpers
- Delete internal/infrastructure/api/xrp/xrplgo/ directory (moved to pkg/chains/xrp/client/)
- Remove obsolete dtoxrp types: ResponseGetAccountInfo, SentTx, TxInfo and nested types
- Regenerate all XRP mocks to reflect updated interface signatures
- Build: zero compilation errors; Tests: 1141 pass, 0 lint issues

* docs: mark all refactoring-chain-rpc spec tasks complete

Update parent group checkboxes (0-5) to [x] after verifying
all sub-tasks are implemented and verified.

* fix(pr): rename ToDTOResponseGetAccountInfo to fromProtoAccountInfo

Rename the converter function to accurately reflect that it converts
a protogen type to a pkg/client type (not a DTO), and update the
stale comment at the call site in xrpapi_account.go.

---------

+++ internal/application/usecase/watch/btc/monitor_transaction.go
@@ -195,7 +195,7 @@ func (u *monitorTransactionUseCase) checkTransactionConfirmation(
 		"required", u.btcClient.ConfirmationBlock())
 
 	// Check if confirmations meet threshold
-	if uint64(tx.Confirmations) >= u.btcClient.ConfirmationBlock() {
+	if tx.Confirmations >= u.btcClient.ConfirmationBlock() {
 		return true, nil
 	}
